requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
transformers>=4.30.0
datasets>=2.14.0
peft>=0.4.0
bitsandbytes>=0.41.0
pydantic>=2.0
python-multipart>=0.0.6